// backend/src/entity/entity-enum.ts
export enum DetailServiceType {
    TC_WORK = 1,
    TEST_WORK = 2,
    TITLE_WORK = 3,
    YOUTUBE_APPLY = 4
}

export enum WorkType {
    TC = 4,
    TRANSLATE = 5,
    TEST = 6
}

export enum Trigger {
    ON="Y",
    OFF="N"
}

export enum SystemLang {
    KOREAN = "KO",
    ENGLISH = "EN",
    CHINESE = "CN",
    JAPANESE = "JP"
}

export enum CurrencyType {  
    JPY = "JPY",
    KRW = "KRW",
    USD = "USD"
}